Update User Account Working
The "update-account.handlebars" has been added and now works correctly with the MySQL database. The user's name is now displayed at the top of "home.handlebars" (it is passed in through the "context" array of the Node route). Additional error handling added for situations that would cause MySQL to throw server terminating exceptions.